fix(variable): fix count of variable error
The following variables were not counting correctly:
- FILES_NAMES
- FILES
- MULTIPART_PART_HEADERS
- REQUEST_COOKIES_NAMES
- REQUEST_COOKIES

+++ src/variable/request_headers.h
@@ -38,7 +38,9 @@ public:
         // collection
         { result.append(static_cast<int>(t.httpExtractor().request_header_count_)); },
         // specify subname
-        { result.append(t.httpExtractor().request_header_find_(sub_name_).size()); });
+        {
+          result.append(static_cast<int>(t.httpExtractor().request_header_find_(sub_name_).size()));
+        });
 
     RETURN_VALUE(
         // collection
